# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/business": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Get authenticated business",
                "description": "Returns the authenticated business details",
                "operationId": "9be6ad8edb737a22e2112453c671d540",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Business Name"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "business@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Business"
                ],
                "summary": "Create a new business",
                "description": "Creates a new business entry",
                "operationId": "1d00ca8538e9223139dd9a09127bc82f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "business_name",
                                    "address",
                                    "city",
                                    "postal_code",
                                    "phone_number",
                                    "website",
                                    "description",
                                    "keywords",
                                    "is_approved"
                                ],
                                "properties": {
                                    "business_name": {
                                        "type": "string",
                                        "example": "My Business"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "My City"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "12345"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "555-5555"
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "http://business.com"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A detailed description"
                                    },
                                    "keywords": {
                                        "type": "string",
                                        "example": "keyword1, keyword2"
                                    },
                                    "is_approved": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Business created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/business/{id}": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Get a specific business by ID",
                "description": "Returns details of a specific business",
                "operationId": "8b9ae0aee76c6942f49074f1f6611c26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business found"
                    },
                    "404": {
                        "description": "Business not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Business"
                ],
                "summary": "Update a specific business",
                "description": "Updates a business record",
                "operationId": "e3e357831413e60b97627525a07dc3f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "business_name",
                                    "address",
                                    "city",
                                    "postal_code",
                                    "phone_number",
                                    "website",
                                    "description",
                                    "keywords",
                                    "is_approved"
                                ],
                                "properties": {
                                    "business_name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "website": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "keywords": {
                                        "type": "string"
                                    },
                                    "is_approved": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/business/{id}/edit": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Retrieve business details for editing",
                "description": "Get business details by ID for editing purposes",
                "operationId": "1f63017c3a0ed6979bda6ec737280ef0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the business to be edited",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BusinessResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business data retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Business not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/business/delete-multiple": {
            "post": {
                "tags": [
                    "Business"
                ],
                "summary": "Delete multiple businesses",
                "description": "Soft delete multiple businesses by IDs",
                "operationId": "6a22186cdf092165fafee11d4d8355d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ids"
                                ],
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Businesses deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "null",
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "3 businesses deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid IDs provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/business/restore-multiple": {
            "post": {
                "tags": [
                    "Business"
                ],
                "summary": "Restore multiple businesses",
                "description": "Restore multiple soft-deleted businesses by IDs",
                "operationId": "06d3e5239a8a986b682e8cff1a8ec475",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ids"
                                ],
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Businesses restored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "null",
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "3 businesses restored successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid IDs provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BusinessResource": {
                "title": "BusinessResource",
                "properties": {
                    "id": {
                        "description": "Business ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Business Name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Business Description",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}